@@ -16,10 +16,9 @@
     "preview": "nuxt preview",
     "postinstall": "nuxt prepare",
     "analyze": "nuxt analyze",
-    "vue-devtools": "vue-devtools",
     "lint": "npm run lint:fix && npm run format",
     "lint:fix": "eslint --fix ./",
-    "format": "prettier --write \"**/*.{vue,js,jsx,cjs,mjs,ts,tsx,cts,mts,html,css,md,mdx,json}\" !public --ignore-path .gitignore"
+    "format": "prettier --write \"**/*.{html,vue,svelte,astro,js,jsx,cjs,mjs,ts,tsx,cts,mts,css,md,mdx,json}\" !public --ignore-path .gitignore"
   },
   "dependencies": {
     "@pinia/nuxt": "^0.5.4",
@@ -30,40 +29,39 @@
     "vuetify": "^3.7.1"
   },
   "devDependencies": {
-    "@iconify/json": "^2.2.242",
+    "@iconify/json": "^2.2.295",
     "@julr/unocss-preset-forms": "^0.1.0",
-    "@nuxt/content": "^2.13.2",
-    "@nuxt/eslint": "^0.3.13",
-    "@nuxt/icon": "^1.5.0",
-    "@nuxt/image": "^1.7.1",
-    "@nuxtjs/critters": "^0.7.2",
-    "@nuxtjs/fontaine": "^0.4.3",
+    "@nuxt/content": "^2.13.4",
+    "@nuxt/eslint": "^0.7.5",
+    "@nuxt/icon": "^1.10.3",
+    "@nuxt/image": "^1.9.0",
+    "@nuxtjs/critters": "^0.8.0",
+    "@nuxtjs/fontaine": "^0.4.4",
     "@nuxtjs/html-validator": "^1.8.2",
     "@nuxtjs/seo": "^2.0.0-rc.19",
-    "@pinegrow/nuxt-module": "^3.0.68",
+    "@pinegrow/nuxt-module": "^3.0.69",
     "@pinegrow/tailwindcss-plugin": "^3.0.35",
     "@pinegrow/vuetify-plugin": "^3.0.54",
     "@resvg/resvg-js": "^2.5.0",
-    "@tailwindcss/forms": "^0.5.8",
-    "@tailwindcss/typography": "^0.5.15",
-    "@types/node": "^22.5.1",
-    "@unocss/nuxt": "^0.62.3",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.16",
+    "@types/node": "^22.10.6",
+    "@unocss/nuxt": "^65.4.0",
     "@unocss/preset-icons": "^0.62.3",
-    "@vue/devtools": "^6.6.3",
-    "@vueuse/nuxt": "^11.0.3",
+    "@vueuse/nuxt": "^12.4.0",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.9.1",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
     "npm-run-all": "^4.1.5",
-    "nuxt": "^3.13.0",
+    "nuxt": "^3.15.1",
     "nuxt-icon": "^0.6.10",
-    "postcss": "^8.4.41",
-    "prettier": "^3.3.3",
+    "postcss": "^8.5.0",
+    "prettier": "^3.4.2",
     "rehype-external-links": "^3.0.0",
     "sass-embedded": "^1.77.8",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^5.5.4",
-    "vuetify-nuxt-module": "^0.18.0"
+    "tailwindcss": "^3.4.17",
+    "typescript": "^5.7.3",
+    "vuetify-nuxt-module": "^0.18.3"
   },
   "pnpm": {
     "overrides": {
