@@ -1,12 +1,27 @@
+import { useHead, useSeoMeta } from '@unhead/vue'
+import { useSchemaOrg, defineWebSite, defineWebPage } from '@unhead/schema-org'
+
 import siteMeta from '@/site'
-// import image from '@/screenshots/image.jpg'
+
+import { heroImageUrl } from '@/utils/hero'
 import { fontUrls } from '@/utils/font'
 
 import checkDarkTheme from '@/composables/dark-color-scheme-check?raw'
 import type { Script } from '@unhead/schema'
 type TurboScript = Script & { once: true }
 
-export const useHeadAndMeta = () => {
+export const useHeadAndMeta = (pageMeta: any) => {
+  const {
+    title: siteTitle,
+    description: siteDescription,
+    url: siteUrl,
+    ogImageUrl: siteOgImageUrl,
+    generator: siteGenerator,
+    author,
+    twitter,
+    titleSeparator,
+  } = siteMeta
+
   const link: any = [
     // ...[
     //   '/fonts/barlow-7cHpv4kjgoGqM7E_Ass52Hs.woff2',
@@ -22,10 +37,23 @@ export const useHeadAndMeta = () => {
     //       href,
     //     } as const),
     // ),
+    {
+      rel: 'icon',
+      // type: 'image/x-icon',
+      type: 'image/svg+xml',
+      href: '/favicon.svg',
+    },
   ]
   const noscript: any = []
 
+  const canonicalUrl = pageMeta.canonicalUrl || siteUrl
+  if (canonicalUrl) {
+    // Canonical URL
+    link.push({ rel: 'canonical', href: canonicalUrl.href })
+  }
+
   if (fontUrls.length) {
+    // Font preloads
     const googleapis = 'https://fonts.googleapis.com'
     const gstatic = 'https://fonts.gstatic.com'
     link.push(
@@ -45,37 +73,24 @@ export const useHeadAndMeta = () => {
     )
   }
 
-  const {
-    title: siteTitle,
-    description: siteDescription,
-    author,
-    twitter,
-    titleSeparator,
-  } = siteMeta
+  const title = pageMeta.title
+    ? `${pageMeta.title} ${titleSeparator} ${siteTitle}`
+    : siteTitle
 
-  const route = useRoute()
-
-  const pageMeta = computed(() => {
-    return {
-      title: route.meta.title as string,
-      description: route.meta.description as string,
-    }
-  })
-
-  const title = computed(() =>
-    pageMeta.value.title
-      ? `${pageMeta.value.title} ${titleSeparator} ${siteTitle}`
-      : siteTitle,
-  )
-
+  // Manage head with useHead
   useHead({
-    title, // title is defined statically using definePageMeta in pages and resolved here. For dynamic routes, this resolved title is further overriden within in the dynamic route itself (for eg, [slug]) using a second useHead.
+    title, // title is either defined statically within the astro page or dynamically by adding it to the page's frontmatter, and it is resolved here.
 
     // Other unused params - titleTemplate, templateParams
     titleTemplate: null,
 
-    // Instead of setting meta via useHead, useServerSeoMeta or @nuxtjs/seo is used.
-    // meta: [],
+    // Instead of setting other meta here, useSeoMeta is used.
+    meta: [
+      {
+        name: 'twitter:url',
+        content: canonicalUrl.href,
+      },
+    ],
 
     // useScript can also be used to load scripts
     script: [{ innerHTML: checkDarkTheme, once: true } as TurboScript],
@@ -86,39 +101,60 @@ export const useHeadAndMeta = () => {
     style: [],
   })
 
-  const description = computed(
-    () => pageMeta.value.description || siteDescription,
-  )
+  const description = pageMeta.description || siteDescription
+  const generator = pageMeta.generator || siteGenerator
+  const keywords = pageMeta.tags?.toString()
 
-  useServerSeoMeta({
+  let siteOgImage: string = siteUrl
+
+  try {
+    siteOgImage = new URL(siteOgImageUrl, siteUrl).href
+  } catch (err) {
+    // console.log(err)
+  }
+
+  const ogImage = pageMeta.ogImage || siteOgImage || heroImageUrl
+
+  if (ogImage.src) {
+    // If pageMeta.ogImage was imported within the app, Vite returns it as an object with src, width, height etc. So, replace src with url
+    ogImage.url = ogImage.src
+    delete ogImage.src
+  }
+
+  // Manage head meta with useSeoMeta
+  useSeoMeta({
     title,
     description,
     author,
-    // charset: 'utf-8', // defaulted by nuxt
-    // viewport: 'width=device-width, initial-scale=1', // defaulted by nuxt
-    // keywords: route.meta.tags?.toString(), // can be ignored, not used anywhere
+    charset: 'utf-8',
+    viewport: 'width=device-width, initial-scale=1',
+    generator,
+    keywords,
 
-    // ogTitle: theTitle, // set by @nuxtjs/seo's nuxt-seo-utils
-    // ogDescription: description, // set by @nuxtjs/seo's nuxt-seo-utils
-    // ogType: 'website', // set by @nuxtjs/seo's nuxt-seo-utils
-    // ogImage: image, // set by @nuxtjs/seo's nuxt-og-image
-    ogImageAlt: title, // set by @nuxtjs/seo's nuxt-og-image
+    // // Open Graph / Facebook / LinkedIn / Discord
+    ogTitle: title,
+    ogDescription: description,
+    ogType: 'website',
+    ogImage: ogImage,
+    ogImageAlt: title,
     // // Other values - og:image:width, og:image:height, og:image:alt, og:image:type, og:image:secure_url
-    // ogUrl: url, // set by @nuxtjs/seo
-    // ogSiteName: theTitle, // set by @nuxtjs/seo
+    ogUrl: canonicalUrl.href,
+    ogSiteName: title,
     // // Other values - og: locale, og: type
 
+    // // Twitter (X)
+    twitterCard: 'summary_large_image',
     twitterTitle: title,
     twitterDescription: description,
-    // twitterImage: image, // set by @nuxtjs/seo & nuxt-og-image
+    twitterImage: ogImage,
     twitterImageAlt: title,
     twitterSite: twitter,
     twitterCreator: '@techakayy',
-    // twitterCard: 'summary_large_image', // set by @nuxtjs/seo & nuxt-og-image
   })
 
+  // Manage schema-org with useSchemaOrg
   // https://unhead.unjs.io/schema-org/getting-started/setup#_3-add-site-schemaorg
-  // https://nuxtseo.com/schema-org/guides/quick-setup
+  // https://nuxtseo.com/learn/mastering-meta/schema-org#reactivity-with-useschemaorg
   useSchemaOrg([
     defineWebSite({
       name: title,
