@@ -1,7 +1,7 @@
-import type { Store, Products } from '~~/types/store'
-import products from './data'
+import type { Store, Product } from '~~/types/store'
+import productsRaw from './data'
 
-products as Products
+const products: Product[] = productsRaw
 
 export default defineEventHandler(async (/*event*/): Promise<Store> => {
   const categories = [...new Set(products.map((product) => product.category))]
